requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
praw==5.3.0
praw-oauth2util==0.3.5
prawcore==0.13.0
requests==2.18.4
six==1.11.0
update-checker==0.16
urllib3==1.22
Werkzeug==0.14.1
